Test for Price Increase message
This should function to read the time zone given from the interview and convert the price change date from ET to the client's time zone. Hopefully works with DST but needs to be tested. Changes message depending on if price changes by a set amount or to a set amount.